fastapi/requirements.txt
fastapi==0.65.2
uvicorn
pillow==8.2.0
torch==1.9.0+cpu
torchvision==0.10.0+cpu
python-multipart==0.0.5 
aiofiles==0.4.0
black==21.6b0
certifi==2021.5.30
chardet==4.0.0
idna==2.10
numpy==1.19.5
pandas==1.3.0
python-dateutil==2.8.1
python-magic==0.4.24
pytz==2021.1
requests==2.25.1
six==1.16.0
tqdm==4.61.2
git+https://github.com/huggingface/transformers.git
urllib3==1.26.6
datasets >= 1.1.3
jax>=0.2.8
jaxlib>=0.1.59
flax>=0.3.4
optax>=0.0.9
sentencepiece
mtranslate